requirements.txt
flask
requests
PyYAML